feature/juan/clean Feat: Update CharacterAdapter
